requirements.txt

Flask
numpy
Pillow
tensorflow
google-generativeai
gunicorn
